+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to the Pony compiler and standard library will be documented
 ### Changed
 
 - Rename MaybePointer to NullablePointer ([PR #3293](https://github.com/ponylang/ponyc/pull/3293))
+- `--ponyminthreads` option can't be larger than `--ponythreads` ([PR #3303](https://github.com/ponylang/ponyc/pull/3303))
 
 ## [0.31.0] - 2019-08-31